+++ tidypolars/tidypolars.py
@@ -115,6 +115,33 @@ class tibble(pl.DataFrame):
         # TODO: Allow to work on multiple inputs
         return self.vstack(df).pipe(as_tibble)
 
+    def distinct(self, *args) -> "tp.tibble":
+        """
+        Select distinct/unique rows
+
+        Parameters
+        ----------
+        *args : Expr
+            Column expressions find distinct/unique rows
+
+        Examples
+        --------
+        df = tp.tibble({'a': range(3), 'b': ['a', 'a', 'b']})
+        
+        df.distinct()
+        df.distinct('b')
+
+        """
+        # TODO: Create for series
+        args = args_as_list(args)
+
+        if len(args) == 0:
+            df = self.drop_duplicates()
+        else:
+            df = self.select(args).drop_duplicates()
+        
+        return df.pipe(as_tibble)
+
     def filter(self, *args) -> "tp.tibble":
         """
         Filter rows on one or more conditions
@@ -126,7 +153,7 @@ class tibble(pl.DataFrame):
 
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': ['a', 'a', 'b'])
+        df = tp.tibble({'a': range(3), 'b': ['a', 'a', 'b']})
         
         df.filter(col('a') < 2, col('c') == 'a')