+++ tests/test_groupby.py
@@ -23,6 +23,13 @@ def test_group_mutate():
     expected = tp.Tibble({'x': [0, 1], 'y': ['a', 'b'], 'avg_x': [0, 1]})
     assert actual.frame_equal(expected), "group mutate failed"
 
+def test_group_slice():
+    """Can slice by group"""
+    df = tp.Tibble({'x': range(3), 'y': ['a', 'a', 'b']})
+    actual = df.slice(0, groupby = 'y').arrange('y')
+    expected = tp.Tibble({'x': [0, 2], 'y': ['a', 'b']})
+    assert actual.frame_equal(expected), "group slice failed"
+
 def test_group_summarize():
     """Can summarize by group"""
     df = tp.Tibble({'x': range(3), 'y': ['a', 'a', 'b']})