+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   + `.bind_cols()`
   + `.bind_rows()`
   + `.pull()`
+  + `.distinct()`
 
 ## v0.1.3 (2021/10/04)